sentry moved to components
